requirements.txt
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
cryptography==43.0.3
distro==1.9.0
dnspython==2.7.0
execnet==2.1.1
filelock==3.16.1
Flask==3.0.3
Flask-WTF==1.2.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jwt==1.3.1
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
MarkupSafe==3.0.1
networkx==3.4.2
numpy==2.1.2
openai==1.51.2
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
pytest==8.3.3
pytest-mock==3.14.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
python-nmap==0.7.1
# pywin32==308
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.0
Werkzeug==3.0.4
WTForms==3.1.2
pytest-asyncio
aihttp==3.11.2
beautifulsoup4==4.12.3